+++ api/index.ts
@@ -1,7 +1,7 @@
 import express from 'express'
 import handleDiscussions from '../src/features/discussions'
 import handleLanguages from '../src/features/languages'
-import { fetcher, handleContributions } from '../src/utils'
+import handleContributions from '../src/features/contributions'
 
 const app = express()
 
@@ -9,12 +9,9 @@ app.get('/api', async (req, res) => {
   const { username, contributions, discussions, languages } = req.query
   if (!username) return res.status(400).send('Please provide a valid username')
 
-  // TODO: refac as const, add as variables.x = y
-  let variables: Record<string, any> = { login: username as string }
   let result = {}
 
   if (contributions) {
-    const query: string = require('../src/queries/contributions').default
     const stars = req.query['contributions.stars'] as string
     const owner = req.query['contributions.owner'] as string
     // Allowed types: org, all
@@ -24,57 +21,15 @@ app.get('/api', async (req, res) => {
     // Queries: commit, issue, pull, repo, review
     const type = req.query['contributions.type'] as string
 
-    if (type) {
-      const contributionTypes = type.split(',').map((el) => {
-        if (el === 'pull') return 'PULL_REQUEST'
-        if (el === 'repo') return 'REPOSITORY'
-        if (el === 'review') return 'PULL_REQUEST_REVIEW'
-
-        return el.toUpperCase()
-      })
-
-      variables = { ...variables, contributionTypes }
-    } else {
-      const contributionTypes = [
-        'COMMIT',
-        'ISSUE',
-        'PULL_REQUEST',
-        'REPOSITORY',
-        'PULL_REQUEST_REVIEW',
-      ]
-
-      variables.contributionTypes = contributionTypes
-    }
-
-    if (owner) variables = { ...variables, owner: true }
-
-    let nodesArray: any[] = []
-    // init loop
-    let hasNextPage = true
-
-    while (hasNextPage) {
-      const {
-        data: {
-          user: {
-            repositoriesContributedTo: {
-              nodes,
-              pageInfo: { hasNextPage: newHasNextPage, endCursor },
-            },
-          },
-        },
-      } = await fetcher(query, variables)
-      nodesArray = [...nodesArray, ...nodes]
-      variables = { ...variables, cursor: endCursor }
-      hasNextPage = newHasNextPage
-
-      const contribs = handleContributions(
-        nodesArray,
-        stars,
-        accountType,
-        ignored,
-      )
-      result = { ...result, ...contribs }
-    }
+    const contribs = await handleContributions({
+      login: username as string,
+      stars,
+      owner,
+      accountType,
+      ignored,
+      type,
+    })
+    result = { ...result, contribs }
   }
 
   if (discussions) {
